+++ examples/ERP/plot_classify_MEG_mdm.py
@@ -1,6 +1,6 @@
 """
 ====================================================================
-Decoding in sensor space data using Riemannian Geometry and XDAWN
+Multiclass MEG ERP Decoding
 ====================================================================
 
 Decoding applied to MEG data in sensor space decomposed using Xdawn.
@@ -11,25 +11,23 @@ classified by the MDM algorithm (Nearest centroid).
 for mean-covariance matrices used by the classification algorithm.
 
 """
-# Authors: Alexandre Gramfort <alexandre.gramfort@telecom-paristech.fr>
-#          Romain Trachel <romain.trachel@inria.fr>
-#          Alexandre Barachant <alexandre.barachant@gmail.com>
+# Authors: Alexandre Barachant <alexandre.barachant@gmail.com>
 #
 # License: BSD (3-clause)
 
 import numpy as np
-from pylab import *
+from matplotlib import pyplot as plt
 from pyriemann.estimation import XdawnCovariances
-from pyriemann.classification import MDM,FgMDM
+from pyriemann.classification import MDM
+from pyriemann.utils.viz import plot_confusion_matrix
 
 import mne
 from mne import io
 from mne.datasets import sample
 
-from sklearn.pipeline import Pipeline  # noqa
+from sklearn.pipeline import make_pipeline
 from sklearn.cross_validation import KFold
-from sklearn.metrics import classification_report,confusion_matrix
-from sklearn.linear_model import LogisticRegression
+from sklearn.metrics import classification_report
 
 print(__doc__)
 
@@ -40,7 +38,7 @@ data_path = sample.data_path()
 raw_fname = data_path + '/MEG/sample/sample_audvis_filt-0-40_raw.fif'
 event_fname = data_path + '/MEG/sample/sample_audvis_filt-0-40_raw-eve.fif'
 tmin, tmax = -0., 1
-event_id = dict(aud_l=1,aud_r=2, vis_l=3,vis_r=4)
+event_id = dict(aud_l=1, aud_r=2, vis_l=3, vis_r=4)
 
 # Setup for reading the raw data
 raw = io.Raw(raw_fname, preload=True)
@@ -53,67 +51,46 @@ picks = mne.pick_types(raw.info, meg='grad', eeg=False, stim=False, eog=False,
 
 # Read epochs
 epochs = mne.Epochs(raw, events, event_id, tmin, tmax, proj=False,
-                    picks=picks, baseline=None, preload=True)
+                    picks=picks, baseline=None, preload=True, verbose=False)
 
 labels = epochs.events[:, -1]
 evoked = epochs.average()
 
 ###############################################################################
-# Decoding in sensor space using a linear SVM
-
+# Decoding with Xdawn + MDM
 
 n_components = 3  # pick some components
 
 # Define a monte-carlo cross-validation generator (reduce variance):
-cv = KFold(len(labels), 10,shuffle=True, random_state=42)
+cv = KFold(len(labels), 10, shuffle=True, random_state=42)
 pr = np.zeros(len(labels))
 epochs_data = epochs.get_data()
 
-
-
 print('Multiclass classification with XDAWN + MDM')
-clf = Pipeline([('COV',XdawnCovariances(n_components)),('MDM',MDM())])
+
+clf = make_pipeline(XdawnCovariances(n_components), MDM())
 
 for train_idx, test_idx in cv:
     y_train, y_test = labels[train_idx], labels[test_idx]
-    
+
     clf.fit(epochs_data[train_idx], y_train)
     pr[test_idx] = clf.predict(epochs_data[test_idx])
 
-print classification_report(labels,pr)
-print confusion_matrix(labels,pr)
+print(classification_report(labels, pr))
 
-print('Multiclass classification with XDAWN + FgMDM')
-clf = Pipeline([('COV',XdawnCovariances(n_components)),('MDM',FgMDM())])
-
-for train_idx, test_idx in cv:
-    y_train, y_test = labels[train_idx], labels[test_idx]
-    
-    clf.fit(epochs_data[train_idx], y_train)
-    pr[test_idx] = clf.predict(epochs_data[test_idx])
-
-print classification_report(labels,pr)
-print confusion_matrix(labels,pr)
-
-# spatial patterns
+###############################################################################
+# plot the spatial patterns
 xd = XdawnCovariances(n_components)
-Cov = xd.fit_transform(epochs_data,labels)
+xd.fit(epochs_data, labels)
 
-evoked.data = xd.Xd._patterns.T
+evoked.data = xd.Xd.patterns_.T
 evoked.times = np.arange(evoked.data.shape[0])
-evoked.plot_topomap(times=[0, n_components, 2*n_components,3*n_components], ch_type='grad',
-                    colorbar=False, size=1.5)
-                    
-# prototyped covariance matrices
-mdm = MDM()
-mdm.fit(Cov,labels)
-fig,axe = plt.subplots(1,4)
-axe[0].matshow(mdm.covmeans[0])
-axe[0].set_title('Class 1 covariance matrix')
-axe[1].matshow(mdm.covmeans[1])
-axe[1].set_title('Class 2 covariance matrix')
-axe[2].matshow(mdm.covmeans[2])
-axe[2].set_title('Class 3 covariance matrix')
-axe[3].matshow(mdm.covmeans[3])
-axe[3].set_title('Class 4 covariance matrix')
+evoked.plot_topomap(times=[0, n_components, 2 * n_components,
+                    3 * n_components], ch_type='grad', colorbar=False,
+                    size=1.5)
+
+###############################################################################
+# plot the confusion matrix
+names = ['audio left', 'audio right', 'vis left', 'vis right']
+plot_confusion_matrix(labels, pr, names)
 plt.show()