+++ pyriemann/utils/viz.py
@@ -1,0 +1,21 @@
+"""Helpers for vizualization."""
+import seaborn as sns
+import pandas as pd
+import numpy as np
+
+from sklearn.metrics import confusion_matrix
+
+
+def plot_confusion_matrix(targets, predictions, target_names,
+                          title='Confusion matrix', cmap="Blues"):
+    """Plot Confusion Matrix."""
+    cm = confusion_matrix(targets, predictions)
+    cm = 100 * cm.astype('float') / cm.sum(axis=1)[:, np.newaxis]
+
+    df = pd.DataFrame(data=cm, columns=target_names, index=target_names)
+    g = sns.heatmap(df, annot=True, fmt=".1f", linewidths=.5, vmin=0, vmax=100,
+                    cmap=cmap)
+    g.set_title(title)
+    g.set_ylabel('True label')
+    g.set_xlabel('Predicted label')
+    return g